Move MetricsNames to types package
